Removed coeff_generator from beamform_op_sequence. General code cleanup. Renamed folder 'test' to 'unit_test' as 'test' seemed to shadow another python system folder and caused import issues. All pytests pass.
