+++ beamformer/beamforming/beamform_op_sequence.py
@@ -7,105 +7,11 @@ per the shape descibed. Provision for batched operations is included, i.e. reord
 passed to the kernel in a single array.
 """
 
-import numpy as np
 from beamform_reorder import prebeamform_reorder
 from beamforming import matrix_multiply
 from katsdpsigproc import accel
 from katsdpsigproc.abc import AbstractContext
-
-# from test import coeff_generator
-
-
-class CoeffGenerator:
-    """Class for generating coefficients for testing purposes.
-
-    Note: The coefficients for this feature will be supplied by CAM and are not
-    generated by CBF. This code can removed when suitable coefficients are available.
-
-    Parameters
-    ----------
-    n_batches: int
-        The number of matrices to be reordered, a single data matrix = one batch.
-    pols: int
-        Number of polarisations. Always 2.
-    n_channels: int
-        The number of frequency channels to be processed.
-    n_blocks: int
-        The number of blocks that each channels set of samples are divided into.
-    samples_per_block: int
-        The number of time samples to be processed per block.
-    n_ants: int
-        The number of antennas that will be used in beamforming. Each antennas is expected to produce two polarisations.
-    """
-
-    def __init__(self, n_batches, pols, n_channels, n_blocks, samples_per_block, n_ants) -> None:
-        """Initialise the coefficient generation class."""
-        self.n_batches = n_batches
-        self.pols = pols
-        self.n_channels = n_channels
-        self.n_blocks = n_blocks
-        self.samples_per_block = samples_per_block
-        self.ants = n_ants
-        self.complexity = 2
-
-    def Coeffs(self, test_id) -> np.ndarray:
-        """Generate coefficients for complex multiplication.
-
-        Note: This is for use in complex multiplication using two
-        real-valued arrays. For this reason the coefficients need to be
-        arranged as follows.
-
-        Coefficients Array:
-        [R00  I00
-         -I00 R00
-         R10  I10
-         -I10 R10
-         ...  ...]
-        Where:  R00 = Real coeff 0
-                I00 = Imag coeff 0
-        and the matrix is structured as a N x 2 array.
-
-        Returns
-        -------
-        np.ndarray of type float.
-        Output array of test coefficients.
-        """
-        length = self.n_batches * self.pols * self.n_channels * self.n_blocks * self.samples_per_block
-
-        # Preform matrix prior to setting values
-        coeffs = np.ones(((self.ants * 2) * 2 * length), np.float32).reshape(length, 2, self.ants * 2)
-
-        # Set static values for testing
-        real_value = 4
-        imag_value = 1
-        for i in range(coeffs.shape[0]):
-            for j in range(coeffs.shape[1]):
-                for k in range(coeffs.shape[2]):
-                    if j == 0:
-                        if k % 2:
-                            coeffs[i, j, k] = -1 * imag_value
-                        else:
-                            coeffs[i, j, k] = real_value
-                    else:
-                        if k % 2:
-                            coeffs[i, j, k] = real_value
-                        else:
-                            coeffs[i, j, k] = imag_value
-
-        # Temporary use to identify which complex multiplication will be used.
-        # This affects the data shape.
-        if test_id == "kernel":
-            return coeffs.reshape(
-                self.n_batches,
-                self.pols,
-                self.n_channels,
-                self.n_blocks,
-                self.samples_per_block,
-                2,
-                self.ants * self.complexity,
-            )
-        else:
-            return coeffs
+from unit_test import coeff_generator
 
 
 class BeamformSeqTemplate:
@@ -237,13 +143,11 @@ if __name__ == "__main__":
     test_id = "kernel"
 
     # Generate coefficients
-    # coeff_gen = coeff_generator.CoeffGenerator(batches, n_channels_per_stream, n_blocks, samples_per_block, n_ants)
-    coeff_gen = CoeffGenerator(batches, pols, n_channels_per_stream, n_blocks, samples_per_block, n_ants)
-    # if test_id == 'kernel':
-    #     coeffs = coeff_gen.GPU_Coeffs_kernel()
-    # elif test_id == 'sgemm':
-    #     coeffs = coeff_gen.GPU_Coeffs_cublas
-    coeffs = coeff_gen.Coeffs(test_id)
+    coeff_gen = coeff_generator.CoeffGenerator(batches, n_channels_per_stream, n_blocks, samples_per_block, n_ants)
+    if test_id == "kernel":
+        coeffs = coeff_gen.GPU_Coeffs_kernel()
+    elif test_id == "sgemm":
+        coeffs = coeff_gen.GPU_Coeffs_cublas
 
     ctx = accel.create_some_context(device_filter=lambda x: x.is_cuda, interactive=False)
     queue = ctx.create_command_queue()
