+++ beamformer/unit_test/__init__.py
@@ -1,0 +1,2 @@
+"""Import the following modules for the unit test package."""
+__all__ = ["coeff_generator", "complex_mult_cpu"]
